hiq/requirements.txt
asn1crypto==0.24.0
beautifulsoup4==4.6.0
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
cryptography==2.3.1
Django==2.0.8
django-extensions==2.1.2
django-modelcluster==4.2
django-taggit==0.23.0
django-treebeard==4.3
djangorestframework==3.8.2
draftjs-exporter==2.1.2
html5lib==1.0.1
idna==2.7
jwt==0.5.4
mysqlclient==1.3.13
Pillow==5.2.0
pycparser==2.19
pytz==2018.5
requests==2.19.1
six==1.11.0
Unidecode==1.0.22
urllib3==1.23
wagtail==2.2.2
webencodings==0.5.1
Werkzeug==0.14.1
Willow==1.1
